fix initializers for options
